+++ authentication/routes/routes.js
@@ -1,10 +1,10 @@
 const { Router } = require("express");
 const router = Router();
-const authenticate = require("../controller/authentication");
-const admin = require("../controller/admin");
-const valid = require("../services/valid");
 
-router.post("/signup", authenticate.postsignup);
-router.post("/login", authenticate.postlogin);
+const signUpController = require("../controller/Normal_User/signUp");
+const loginController = require("../controller/Normal_User/login");
+
+router.post("/signup", signUpController.postsignup);
+router.post("/login", loginController.postlogin);
 
 module.exports = router;
